packages/eval-harness/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "loop-eval"
version = "0.1.0"
description = "Loop eval harness: scorers, datasets, runner."
license = { text = "Apache-2.0" }
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2",
    "pyyaml>=6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["loop_eval"]
